requirements.txt
python-telegram-bot==20.0a1
google-generativeai==0.1.1
pg8000==1.20
bcrypt==2.0.0